feat: add isExpired helper
I found myself sometimes needing to manually check the validity/expired-stat of a cache entry
the exported `shouldRefresh` helper is capable of doing that but
the naming is unintuitive when used outside of cachified itself

when I look at a cache entry manually, i want to know "is is expired/valid" not "should i refresh"
